% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCutoff.R
\name{PlotCutoff}
\alias{PlotCutoff}
\title{Plot Cutoff Lines onto ECDF Plot}
\usage{
PlotCutoff(h = NULL, v = NULL, vectorName = 1)
}
\arguments{
\item{h}{Numeric value indicating position of horizontal line}

\item{v}{Numeric value indicating position of vertical line}

\item{vectorName}{Name (string) of curve of interest for which either a
horizontal or vertical threshold is to be determined. The name is derived
from the original name given to relevant vectors when PlotMultiECDFs was
called. If not specified, the curve corresponding to the vector 
positioned first in the aforemention list is used.}
}
\description{
Plots a set of vertical and horizontal cutoff lines given either a starting
horizontal or vertical position and a curve of interest. The point at which
the relevant line interesects the y or x-axis is indicated. A horizontal
starting line position is used as default. Multiple line sets can be
plotted onto the same plot
}
\examples{
Add_Threshold_Line(v=0.5, vectorName="a")
Add_Threshold_Line(h=0.3)
Add_Threshold_Line(0.2)
}
