% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMultiECDFs.R
\name{PlotMultiECDFs}
\alias{PlotMultiECDFs}
\title{Plot Multiple Vectors as ECDFs of the Same Plot}
\usage{
PlotMultiECDFs(x, colours = rainbow(n = length(x)), ...)
}
\arguments{
\item{x}{A named list of double-precision vectors}

\item{colours}{A character vector to determine plotted curve colours. Default colours are provided}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Plots multiple ECDFs onto same plot using named list of vectors as input.
}
\examples{
example_vectors = list(a=rnorm(1000), b=rnorm(1000), c=rnorm(1000, 
 mean=0.3))
PlotMultiECDFs (example_vectors, xlim=range(-3, 3), main='foo', xlab='bar')
}
